backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai>=0.8.0
firebase-admin==6.5.0
python-dotenv==1.0.1
httpx==0.27.0
openai>=1.50.0
